Add hooks to Presenter, clean up redundant methods
Most logic in Presenters was placed in the load or loadView methods,
adding repetitive calls to loadView and checks to the initialization
state of the view. To DRY things up, add hooks into various stages of
Presenter's life: beforeLoad, beforeInitialize, afterInitialize,
beforeViewLoad, and beforeUnload.

These hooks make it much easier to customize each presenter and allow me
to delete new declarations of load and loadView.