Each Presenter only creates one ItemsPresenter
Previously, each presenter initiated a new ItemsPresenter on load to
pass a new set of items. This made it impossible to store state in the
presenter, namely its sorting options.

Instead, now only initiate one ItemsPresenter and pass it the parent
presenter's items array. The array is updated on each load cycle, and
ItemsPresenter can access it.

The array must not be deleted or reassigned or the ItemsPresenter will
no longer show updated items. May update this in the future to make it
more robust.