Move showComplete logic to ItemsPresenter
ItemsPresenter is now in charge of showComplete logic. Instead of
filtering out completed items at the parent presenter, ItemsPresenter
checks its showComplete state and the items' isComplete state, then
decides whether or not to hide the item in the list.

ItemPresenter now takes an object of options, which are attached to its
viewProps. The hide property tells the view to add a class to the item's
container to set its display to none.