Allow Deletion of Projects and Items
Items and projects can now be deleted through the ui. Changes include:
creation of ModalConfirmationPresenter,  changes to ModalView to
accommodate it, and a change to the presenter's unload method.

ModalConfirmationPresenter takes a function for when a user confirms an
action and optionally for when they cancel an action.

Presenter's unload function now checks the view's isInitialized function before trying to
remove the view. This safeguards the presenter from errors if unload is
called by itself and later by a higher order presenter (such as
AppPresenter).

Would like to consider ways for Presenters to request new Presenters
to replace them.