Fix load error in ProjectListPresenter
beforeLoad algorithm assumed that projects were never deleted, so it
didn't check that they still existed in the DB when reloading. Now, the
project-wrapping object contains a boolean value representing whether it
has been confirmed in the DB or not. Any project that remains
unconfirmed after the DB check will be removed from the list of
projects.

All boolean values are reset after load. To accomplish this, a new hook
was added to the Presenter object: afterLoad.