Flesh out ItemDetailedView
Copied in styles from ProjectView, added styling to notes and item
details (date and property) as well as adding delete buttons for each
note.

Created a temporary breadcrumb that displays the project that an item is
a part of, or "All Items" if the item has none. Need to figure out how
to properly route in this one-page app. Then I may recreate the
breadcrumb as a separate presenter.